+++ src/vite-env.d.ts
@@ -28,5 +28,6 @@ declare module 'ionicons/icons' {
   export const homeOutline: string;
   export const scaleOutline: string;
   export const bedOutline: string;
+  export const personCircleOutline: string;
   export default icons;
 }